requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
sqlparse==0.4.3
